More more terminal zoom code into terminalContrib
