+++ src/vs/platform/terminal/common/terminal.ts
@@ -116,7 +116,6 @@ export const enum TerminalSettingId {
 	SmoothScrolling = 'terminal.integrated.smoothScrolling',
 	IgnoreBracketedPasteMode = 'terminal.integrated.ignoreBracketedPasteMode',
 	FocusAfterRun = 'terminal.integrated.focusAfterRun',
-	MouseWheelZoom = 'terminal.integrated.mouseWheelZoom',
 
 	// Debug settings that are hidden from user
