+++ src/vs/workbench/contrib/terminalContrib/zoom/browser/terminal.zoom.contribution.ts
@@ -13,12 +13,13 @@ import { TerminalSettingId } from 'vs/platform/terminal/common/terminal';
 import { IDetachedTerminalInstance, ITerminalContribution, ITerminalInstance, IXtermTerminal } from 'vs/workbench/contrib/terminal/browser/terminal';
 import { registerTerminalContribution } from 'vs/workbench/contrib/terminal/browser/terminalExtensions';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
-import { ITerminalProcessInfo, ITerminalProcessManager, TerminalCommandId } from 'vs/workbench/contrib/terminal/common/terminal';
+import { ITerminalProcessInfo, ITerminalProcessManager } from 'vs/workbench/contrib/terminal/common/terminal';
 import { TerminalWidgetManager } from 'vs/workbench/contrib/terminal/browser/widgets/widgetManager';
 import { registerTerminalAction } from 'vs/workbench/contrib/terminal/browser/terminalActions';
 import { localize2 } from 'vs/nls';
 import { isNumber } from 'vs/base/common/types';
 import { defaultTerminalFontSize } from 'vs/workbench/contrib/terminal/common/terminalConfiguration';
+import { TerminalZoomCommandId, TerminalZoomSettingId } from 'vs/workbench/contrib/terminalContrib/zoom/common/terminal.zoom';
 
 class TerminalMouseWheelZoomContribution extends Disposable implements ITerminalContribution {
 	static readonly ID = 'terminal.mouseWheelZoom';
@@ -46,8 +47,8 @@ class TerminalMouseWheelZoomContribution extends Disposable implements ITerminal
 
 	xtermOpen(xterm: IXtermTerminal & { raw: RawXtermTerminal }): void {
 		this._register(Event.runAndSubscribe(this._configurationService.onDidChangeConfiguration, e => {
-			if (!e || e.affectsConfiguration(TerminalSettingId.MouseWheelZoom)) {
-				if (!!this._configurationService.getValue(TerminalSettingId.MouseWheelZoom)) {
+			if (!e || e.affectsConfiguration(TerminalZoomSettingId.MouseWheelZoom)) {
+				if (!!this._configurationService.getValue(TerminalZoomSettingId.MouseWheelZoom)) {
 					this._setupMouseWheelZoomListener(xterm.raw);
 				} else {
 					this._listener.clear();
@@ -125,7 +126,7 @@ class TerminalMouseWheelZoomContribution extends Disposable implements ITerminal
 registerTerminalContribution(TerminalMouseWheelZoomContribution.ID, TerminalMouseWheelZoomContribution, true);
 
 registerTerminalAction({
-	id: TerminalCommandId.FontZoomIn,
+	id: TerminalZoomCommandId.FontZoomIn,
 	title: localize2('fontZoomIn', 'Increase Font Size'),
 	run: async (c, accessor) => {
 		const configurationService = accessor.get(IConfigurationService);
@@ -137,7 +138,7 @@ registerTerminalAction({
 });
 
 registerTerminalAction({
-	id: TerminalCommandId.FontZoomOut,
+	id: TerminalZoomCommandId.FontZoomOut,
 	title: localize2('fontZoomOut', 'Decrease Font Size'),
 	run: async (c, accessor) => {
 		const configurationService = accessor.get(IConfigurationService);
@@ -149,7 +150,7 @@ registerTerminalAction({
 });
 
 registerTerminalAction({
-	id: TerminalCommandId.FontZoomReset,
+	id: TerminalZoomCommandId.FontZoomReset,
 	title: localize2('fontZoomReset', 'Reset Font Size'),
 	run: async (c, accessor) => {
 		const configurationService = accessor.get(IConfigurationService);
