+++ src/vs/workbench/contrib/terminalContrib/zoom/common/terminal.zoom.ts
@@ -1,0 +1,29 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Microsoft Corporation. All rights reserved.
+ *  Licensed under the MIT License. See License.txt in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+
+import type { IStringDictionary } from 'vs/base/common/collections';
+import { isMacintosh } from 'vs/base/common/platform';
+import { localize } from 'vs/nls';
+import type { IConfigurationPropertySchema } from 'vs/platform/configuration/common/configurationRegistry';
+
+export const enum TerminalZoomCommandId {
+	FontZoomIn = 'workbench.action.terminal.fontZoomIn',
+	FontZoomOut = 'workbench.action.terminal.fontZoomOut',
+	FontZoomReset = 'workbench.action.terminal.fontZoomReset',
+}
+
+export const enum TerminalZoomSettingId {
+	MouseWheelZoom = 'terminal.integrated.mouseWheelZoom',
+}
+
+export const terminalZoomConfiguration: IStringDictionary<IConfigurationPropertySchema> = {
+	[TerminalZoomSettingId.MouseWheelZoom]: {
+		markdownDescription: isMacintosh
+			? localize('terminal.integrated.mouseWheelZoom.mac', "Zoom the font of the terminal when using mouse wheel and holding `Cmd`.")
+			: localize('terminal.integrated.mouseWheelZoom', "Zoom the font of the terminal when using mouse wheel and holding `Ctrl`."),
+		type: 'boolean',
+		default: false
+	},
+};
